Improve eth mock-swarm StateBackendTest trait
StateBackendTest partially tests execution logic for nonce tracking but
in a way that depends on the user to do some of the calculation for it.
This change aims to improve the situation by changing the function
signature for ledger_propose to also accept the block transactions so
that the StateBackendTest implemenation can fully validate the nonces.

StateBackendTest is also renamed to MockExecution as that is the actual
purpose it serves.

An AccountState struct is also added to InMemory state backend so that
it can be extended in the future to also track and validate balances.